requirements.txt
pip
setuptools
pillow
numpy
img2pdf